made some change, DetailsActivity to DetailsFragment
