+++ .idea/misc.xml
@@ -6,4 +6,11 @@
   <component name="ProjectType">
     <option name="id" value="Android" />
   </component>
+  <component name="VisualizationToolProject">
+    <option name="state">
+      <ProjectState>
+        <option name="scale" value="0.2392626953125" />
+      </ProjectState>
+    </option>
+  </component>
 </project>
