+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApplication"
 
+
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
