--- app/src/main/java/com/example/newFilmSearch/DetailsActivity.kt
@@ -1,39 +1,0 @@
-package com.example.myapplication
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.myapplication.databinding.ActivityDetailsBinding
-import com.example.newFilmSearch.Film
-
-class DetailsActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityDetailsBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_details)
-        binding = ActivityDetailsBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
-
-        //Устанавливаем тулбар
-        setSupportActionBar(binding.detailsToolbar)
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        supportActionBar?.setDisplayShowHomeEnabled(true)
-        binding.detailsToolbar.setNavigationOnClickListener {
-            onBackPressed()
-        }
-
-        setFilmsDetails()
-    }
-
-    private fun setFilmsDetails() {
-        //Получаем наш фильм из переданного бандла
-        val film = intent.extras?.get("film") as Film
-        binding.detailsToolbar.title = film.title
-
-        binding.toolbarLayout.title = film.title
-        binding.detailsDescription.text = film.description
-        binding.detailsPoster.setImageResource(film.poster)
-        binding.detailsToolbar.title = film.title
-
-    }
-}
