+++ app/src/main/java/com/example/newFilmSearch/MainActivity.kt
@@ -1,140 +1,89 @@
 package com.example.myapplication
 
-import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.myapplication.databinding.ActivityMainBinding
 import com.example.newFilmSearch.Film
-import com.example.newFilmSearch.FilmListRecyclerAdapter
-import com.example.newFilmSearch.TopSpacingItemDecoration
+import com.example.newFilmSearch.HomeFragment
 
 class MainActivity : AppCompatActivity() {
-    private lateinit var filmsAdapter: FilmListRecyclerAdapter
     private lateinit var binding: ActivityMainBinding
-    val filmsDataBase = listOf(
-        Film(
-            "The Shawshank Redemption",
-            R.drawable.shawshank,
-            "Two imprisoned men bond over a number of years, finding solace and eventual redemption through acts of common decency."
-        ),
-        Film(
-            "The Godfather",
-            R.drawable.god_father,
-            "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son."
-        ),
-        Film(
-            "The Dark Knight",
-            R.drawable.dark_knight,
-            "When the menace known as the Joker wreaks havoc and chaos on the people of Gotham, Batman must accept one of the greatest psychological and physical tests of his ability to fight injustice."
-        ),
-        Film(
-            "Pulp Fiction",
-            R.drawable.pulp,
-            "The lives of two mob hitmen, a boxer, a gangster and his wife, and a pair of diner bandits intertwine in four tales of violence and redemption."
-        ),
-        Film(
-            "Inception",
-            R.drawable.inception,
-            "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O."
-        ),
-        Film(
-            "Hamilton",
-            R.drawable.hamilton,
-            "The real life of one of America's foremost founding fathers and first Secretary of the Treasury, Alexander Hamilton. Captured live on Broadway from the Richard Rodgers Theater with the original Broadway cast."
-        ),
-        Film(
-            "Gisaengchung",
-            R.drawable.parasites,
-            "Greed and class discrimination threaten the newly formed symbiotic relationship between the wealthy Park family and the destitute Kim clan."
-        ),
-        Film(
-            "Interstellar",
-            R.drawable.interstellar,
-            "A team of explorers travel through a wormhole in space in an attempt to ensure humanity's survival."
-        ),
-        Film(
-            "Joker",
-            R.drawable.joker,
-            "In Gotham City, mentally troubled comedian Arthur Fleck is disregarded and mistreated by society. He then embarks on a downward spiral of revolution and bloody crime. This path brings him face-to-face with his alter-ego: the Joker."
-        ),
-        Film(
-            "1917",
-            R.drawable.seventy,
-            "April 6th, 1917. As a regiment assembles to wage war deep in enemy territory, two soldiers are assigned to race against time and deliver a message that will stop 1,600 men from walking straight into a deadly trap."
-        )
 
-    )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
+        setContentView(binding.root)
         enableEdgeToEdge()
-
         initNavigation()
+        initFragment()
+    }
 
-        //находим наш RV
-        binding.mainRecycler.apply {
-            filmsAdapter =
-                FilmListRecyclerAdapter(object : FilmListRecyclerAdapter.OnItemClickListener {
-                    override fun click(film: Film) {
-                        //Создаем бандл и кладем туда объект с данными фильма
-                        val bundle = Bundle()
-                        //Первым параметром указывается ключ, по которому потом будем искать, вторым сам
-                        //передаваемы объект
-                        bundle.putParcelable("film", film)
-                        //Запускаем наше активити
-                        val intent = Intent(this@MainActivity, DetailsActivity::class.java)
-                        //Прикрепляем бандл к интенту
-                        intent.putExtras(bundle)
-                        //Запускаем активити через интент
-                        startActivity(intent)
-                    }
-                })
-            //Присваиваем адаптер
-            adapter = filmsAdapter
-            //Присвои layoutmanager
-            layoutManager = LinearLayoutManager(this@MainActivity)
-            //Применяем декоратор для отступов
-            val decorator = TopSpacingItemDecoration(8)
-            addItemDecoration(decorator)
-        }
-        //Кладем нашу БД в RV
-        filmsAdapter.addItems(filmsDataBase)
+    private fun initFragment() {
+        supportFragmentManager
+            .beginTransaction()
+            .add(R.id.fragment_placeholder, HomeFragment())
+            .addToBackStack(null)
+            .commit()
+    }
+
+    fun launchDetailsFragment(film: Film) {
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.fragment_placeholder, DetailsFragment())
+            .addToBackStack(null)
+            .commit()
     }
 
     private fun initNavigation() {
-      binding.topAppBar.setOnMenuItemClickListener  {
+        binding.topAppBar.setOnMenuItemClickListener {
             when (it.itemId) {
                 R.id.settings -> {
                     Toast.makeText(this, "Настройки", Toast.LENGTH_SHORT).show()
                     true
                 }
+
                 else -> false
             }
         }
-
-       binding.bottomNavigation.setOnNavigationItemSelectedListener {
+        binding.bottomNavigation.setOnNavigationItemSelectedListener {
 
             when (it.itemId) {
                 R.id.favorites -> {
                     Toast.makeText(this, "Избранное", Toast.LENGTH_SHORT).show()
                     true
                 }
+
                 R.id.watch_later -> {
                     Toast.makeText(this, "Посмотреть похже", Toast.LENGTH_SHORT).show()
                     true
                 }
+
                 R.id.selections -> {
                     Toast.makeText(this, "Подборки", Toast.LENGTH_SHORT).show()
                     true
                 }
+
                 else -> false
             }
+        }
+        fun launchDetailsFragment(film: Film) {
+            //Создаем "посылку"
+            val bundle = Bundle()
+            //Кладем наш фильм в "посылку"
+            bundle.putParcelable("film", film)
+            //Кладем фрагмент с деталями в перменную
+            val fragment = DetailsFragment()
+            //Прикрепляем нашу "посылку" к фрагменту
+            fragment.arguments = bundle
+
+            //Запускаем фрагмент
+            supportFragmentManager
+                .beginTransaction()
+                .replace(R.id.fragment_placeholder, fragment)
+                .addToBackStack(null)
+                .commit()
         }
     }
 }
