+++ build.gradle.kts
@@ -1,6 +1,6 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id("com.android.application") version "8.10.1" apply false
+    id("com.android.application") version "8.11.0" apply false
     id("org.jetbrains.kotlin.android") version "1.8.0" apply false
-    id("com.android.library") version "8.10.1" apply false
+    id("com.android.library") version "8.11.0" apply false
 }
